requirements.txt
paramiko
pandas
mysql-connector-python
python-dotenv
openpyxl
termcolor